+++ src/main/kotlin/com/github/bobi/aemgroovyconsoleplugin/config/ui/AemServerEditDialog.kt
@@ -1,11 +1,19 @@
 package com.github.bobi.aemgroovyconsoleplugin.config.ui
 
+import com.github.bobi.aemgroovyconsoleplugin.services.http.AemServerCredentials
+import com.github.bobi.aemgroovyconsoleplugin.services.http.AemServerHttpConfig
 import com.github.bobi.aemgroovyconsoleplugin.services.http.GroovyConsoleHttpService
 import com.github.bobi.aemgroovyconsoleplugin.services.http.model.GroovyConsoleOutput
+import com.github.bobi.aemgroovyconsoleplugin.services.model.AemCertificateToken
+import com.github.bobi.aemgroovyconsoleplugin.services.model.AemDevToken
+import com.github.bobi.aemgroovyconsoleplugin.services.model.AuthType
+import com.google.gson.Gson
+import com.google.gson.GsonBuilder
 import com.intellij.openapi.application.runInEdt
 import com.intellij.openapi.diagnostic.thisLogger
 import com.intellij.openapi.progress.runModalTask
 import com.intellij.openapi.project.Project
+import com.intellij.openapi.ui.ComboBox
 import com.intellij.openapi.ui.DialogWrapper
 import com.intellij.openapi.ui.ValidationInfo
 import com.intellij.openapi.util.text.HtmlBuilder
@@ -16,13 +24,14 @@ import com.intellij.ui.components.JBCheckBox
 import com.intellij.ui.components.JBTextField
 import com.intellij.ui.dsl.builder.*
 import com.intellij.ui.layout.ValidationInfoBuilder
+import com.intellij.ui.layout.selectedValueIs
 import com.intellij.util.ui.UIUtil
+import org.apache.commons.lang3.exception.ExceptionUtils
 import org.jetbrains.concurrency.AsyncPromise
 import java.awt.event.ActionEvent
-import javax.swing.Action
-import javax.swing.JComponent
-import javax.swing.JLabel
-import javax.swing.JPasswordField
+import javax.swing.*
+
+private const val TOKEN_DESCRIPTION = "Paste JSON from Adobe AEM Developer Console"
 
 /**
  * User: Andrey Bardashevsky
@@ -33,6 +42,8 @@ class AemServerEditDialog(private val project: Project, private val tableItem: A
 
     private val httpService = GroovyConsoleHttpService.getInstance(project)
 
+    private val gson: Gson = GsonBuilder().create()
+
     private val urlRegex = Regex(
         "https?://(www\\.)?[-a-zA-Z0-9@:%._+~#=]{1,256}(\\.[a-zA-Z0-9()]{1,6})?\\b([-a-zA-Z0-9()@:%_+.~#?&/=]*)"
     )
@@ -41,7 +52,13 @@ class AemServerEditDialog(private val project: Project, private val tableItem: A
 
     private lateinit var userField: JBTextField
 
+    private lateinit var authTypeField: ComboBox<AuthType>
+
     private lateinit var passwordField: JPasswordField
+
+    private lateinit var devTokenField: JTextArea
+
+    private lateinit var imsCertField: JTextArea
 
     private lateinit var testResultField: JLabel
 
@@ -76,15 +93,28 @@ class AemServerEditDialog(private val project: Project, private val tableItem: A
                 .component
         }
 
+        row("Authentication Type: ") {
+            authTypeField = comboBox(AuthType.values().toList())
+                .align(AlignX.FILL)
+                .resizableColumn()
+                .bindItem({ tableItem.authType }, { tableItem.authType = it!! })
+                .onChanged { doValidateAll() }
+                .component
+        }
+
         row("Username: ") {
             userField = textField()
                 .align(AlignX.FILL)
                 .resizableColumn()
-                .bindText({ tableItem.user }, { tableItem.user = it })
-                .validationOnInput { validateNotEmpty(it.text) }
-                .validationOnApply { validateNotEmpty(it.text) }
+                .bindText(
+                    { if (tableItem.authType == AuthType.BASIC) tableItem.user else "" },
+                    { if (tableItem.authType == AuthType.BASIC) tableItem.user = it }
+                )
+                .validationOnInput { validateBasicAuthNotEmpty(authTypeField.item, it.text) }
+                .validationOnApply { validateBasicAuthNotEmpty(authTypeField.item, it.text) }
+                .enabledIf(authTypeField.selectedValueIs(AuthType.BASIC))
                 .component
-        }
+        }.visibleIf(authTypeField.selectedValueIs(AuthType.BASIC))
 
         row("Password: ") {
             passwordField = cell(JPasswordField())
@@ -94,20 +124,65 @@ class AemServerEditDialog(private val project: Project, private val tableItem: A
                 .bind(
                     JPasswordField::getPassword,
                     { field, value -> field.text = String(value) },
-                    MutableProperty({ tableItem.password.toCharArray() }, { tableItem.password = String(it) })
+                    MutableProperty(
+                        { if (tableItem.authType == AuthType.BASIC) tableItem.password.toCharArray() else "".toCharArray() },
+                        { if (authTypeField.item == AuthType.BASIC) tableItem.password = String(it) }
+                    )
                 )
-                .validationOnInput { validateNotEmpty(String(it.password)) }
-                .validationOnApply { validateNotEmpty(String(it.password)) }
+                .validationOnInput { validateBasicAuthNotEmpty(authTypeField.item, String(it.password)) }
+                .validationOnApply { validateBasicAuthNotEmpty(authTypeField.item, String(it.password)) }
+                .enabledIf(authTypeField.selectedValueIs(AuthType.BASIC))
                 .component
-        }
+        }.visibleIf(authTypeField.selectedValueIs(AuthType.BASIC))
 
-        row {
-            distributedExecutionField = checkBox("Enable distributed execution")
-                .comment("OSGI 'distributedExecutionEnabled' property must be set to 'true'.", MAX_LINE_LENGTH_NO_WRAP)
+        row("Token: ") {
+            devTokenField = textArea()
+                .rows(3)
                 .align(AlignX.FILL)
                 .resizableColumn()
+                .bindText(
+                    { if (tableItem.authType == AuthType.DEV_TOKEN) tableItem.password else "" },
+                    {
+                        if (authTypeField.item == AuthType.DEV_TOKEN) {
+                            tableItem.user = authTypeField.item.shortTitle
+                            tableItem.password = it
+                        }
+                    }
+                )
+                .validationOnInput { validateDevTokenValidJson(authTypeField.item, it.text) }
+                .validationOnApply { validateDevTokenValidJson(authTypeField.item, it.text) }
+                .enabledIf(authTypeField.selectedValueIs(AuthType.DEV_TOKEN))
+                .component
+            contextHelp(TOKEN_DESCRIPTION)
+        }.visibleIf(authTypeField.selectedValueIs(AuthType.DEV_TOKEN))
+
+        row("Certificate:") {
+            imsCertField = textArea()
+                .rows(3)
+                .align(AlignX.FILL)
+                .resizableColumn()
+                .bindText(
+                    { if (tableItem.authType == AuthType.IMS_CERT) tableItem.password else "" },
+                    {
+                        if (authTypeField.item == AuthType.IMS_CERT) {
+                            tableItem.user = authTypeField.item.shortTitle
+                            tableItem.password = it
+                        }
+                    }
+                )
+                .validationOnInput { validateImsCertValidJson(authTypeField.item, it.text) }
+                .validationOnApply { validateImsCertValidJson(authTypeField.item, it.text) }
+                .enabledIf(authTypeField.selectedValueIs(AuthType.IMS_CERT))
+                .component
+            contextHelp(TOKEN_DESCRIPTION)
+        }.visibleIf(authTypeField.selectedValueIs(AuthType.IMS_CERT))
+
+        row("") {
+            distributedExecutionField = checkBox("Enable distributed execution")
+                .align(AlignX.LEFT)
                 .bindSelected({ tableItem.distributedExecution }, { tableItem.distributedExecution = it })
                 .component
+            contextHelp("OSGI 'distributedExecutionEnabled' property must be set to 'true'.")
         }
 
         row {
@@ -158,28 +233,31 @@ class AemServerEditDialog(private val project: Project, private val tableItem: A
                         runInEdt { updateTestResult("Test fail. Output: ${out.output}", true) }
                     }
                 }.onError { ex ->
-                    runInEdt { updateTestResult("Test fail. Error: ${ex.localizedMessage}", true) }
+                    runInEdt {
+                        val rootCause = ExceptionUtils.getRootCause(ex)
+                        val message = rootCause.localizedMessage ?: rootCause.javaClass.simpleName
+
+                        updateTestResult("Test fail. Error: $message", true)
+                    }
                 }
             }
 
-            val req = object {
-                val url = urlField.text
-                val user = userField.text
-                val password = String(passwordField.password)
-                val script = "print('test')".toByteArray()
-            }
-
+            val config = AemServerHttpConfig(
+                url = urlField.text,
+                authType = authTypeField.item,
+                credentials = AemServerCredentials(
+                    user = userField.text,
+                    password = when (authTypeField.item) {
+                        AuthType.BASIC -> String(passwordField.password)
+                        AuthType.DEV_TOKEN -> devTokenField.text
+                        AuthType.IMS_CERT -> imsCertField.text
+                        else -> ""
+                    }
+                ),
+            )
             runModalTask("Checking AEM Server Connection", project, false) {
                 try {
-                    promise.setResult(
-                        httpService.execute(
-                            req.url,
-                            req.user,
-                            req.password,
-                            req.script,
-                            GroovyConsoleHttpService.Action.EXECUTE
-                        )
-                    )
+                    promise.setResult(httpService.execute(config, "print('test')".toByteArray()))
                 } catch (th: Throwable) {
                     thisLogger().info(th)
                     promise.setError(th)
@@ -205,6 +283,67 @@ class AemServerEditDialog(private val project: Project, private val tableItem: A
 
         testResultField.preferredSize = preferredSize
         testResultField.isVisible = text.isNotBlank()
+    }
+
+    private fun ValidationInfoBuilder.validateBasicAuthNotEmpty(authType: AuthType, value: String): ValidationInfo? {
+        if (authType == AuthType.BASIC) {
+            return validateNotEmpty(value)
+        }
+
+        return null
+    }
+
+    private fun ValidationInfoBuilder.validateDevTokenValidJson(authType: AuthType, value: String): ValidationInfo? {
+        if (authType == AuthType.DEV_TOKEN) {
+            val validationInfo = validateNotEmpty(value)
+
+            if (validationInfo != null) {
+                return validationInfo
+            }
+
+            try {
+                val (ok, statusCode, accessToken) = gson.fromJson(value, AemDevToken::class.java)
+
+                if (!ok || statusCode != 200 || accessToken.isBlank()) {
+                    return error("Invalid json")
+                }
+            } catch (e: Throwable) {
+                return error("Invalid json")
+            }
+        }
+
+        return null
+    }
+
+    private fun ValidationInfoBuilder.validateImsCertValidJson(authType: AuthType, value: String): ValidationInfo? {
+        if (authType == AuthType.IMS_CERT) {
+            val validationInfo = validateNotEmpty(value)
+
+            if (validationInfo != null) {
+                return validationInfo
+            }
+
+            try {
+                val (ok, statusCode, integration) = gson.fromJson(value, AemCertificateToken::class.java)
+
+                if (!ok || statusCode != 200
+                    || integration.imsEndpoint.isBlank()
+                    || integration.org.isBlank()
+                    || integration.id.isBlank()
+                    || integration.technicalAccount.clientId.isBlank()
+                    || integration.technicalAccount.clientSecret.isBlank()
+                    || integration.metascopes.isBlank()
+                    || integration.privateKey.isBlank()
+                    || integration.publicKey.isBlank()
+                ) {
+                    return error("Invalid json")
+                }
+            } catch (e: Throwable) {
+                return error("Invalid json")
+            }
+        }
+
+        return null
     }
 
     private fun ValidationInfoBuilder.validateNotEmpty(value: String): ValidationInfo? {