+++ src/main/kotlin/com/github/bobi/aemgroovyconsoleplugin/services/PasswordsService.kt
@@ -6,6 +6,8 @@ import com.intellij.credentialStore.generateServiceName
 import com.intellij.ide.passwordSafe.PasswordSafe
 
 
+private const val AEM_ACCESS_TOKEN_SUFFIX = "-aem-access-token"
+
 /**
  * User: Andrey Bardashevsky
  * Date/Time: 29.07.2022 22:41
@@ -14,16 +16,28 @@ object PasswordsService {
     private const val AEM_GROOVY_CONSOLE = "AEMGroovyConsole"
 
     fun removeCredentials(id: Long) {
-        PasswordSafe.instance.set(credentialAttributes(id), null)
+        val passwordSafe = PasswordSafe.instance
+        passwordSafe.set(credentialAttributes(id), null)
+        passwordSafe.set(credentialAttributes(id, AEM_ACCESS_TOKEN_SUFFIX), null)
     }
 
     fun setCredentials(id: Long, user: String, password: String) {
-        PasswordSafe.instance.set(credentialAttributes(id), Credentials(user, password))
+        val passwordSafe = PasswordSafe.instance
+
+        passwordSafe.set(credentialAttributes(id), Credentials(user, password))
+        passwordSafe.set(credentialAttributes(id, AEM_ACCESS_TOKEN_SUFFIX), null)
     }
 
     fun getCredentials(id: Long): Credentials? = PasswordSafe.instance.get(credentialAttributes(id))
 
-    private fun credentialAttributes(id: Long) = CredentialAttributes(
-        generateServiceName(AEM_GROOVY_CONSOLE, id.toString()),
+    fun setAccessToken(id: Long, token: String) {
+        PasswordSafe.instance.set(credentialAttributes(id, AEM_ACCESS_TOKEN_SUFFIX), Credentials("token", token))
+    }
+
+    fun getAccessToken(id: Long): String? =
+        PasswordSafe.instance.get(credentialAttributes(id, AEM_ACCESS_TOKEN_SUFFIX))?.getPasswordAsString()
+
+    private fun credentialAttributes(id: Long, suffix: String = "") = CredentialAttributes(
+        generateServiceName(AEM_GROOVY_CONSOLE, id.toString() + suffix),
     )
 }