+++ src/main/kotlin/com/github/bobi/aemgroovyconsoleplugin/services/http/GroovyConsoleHttpService.kt
@@ -3,7 +3,9 @@ package com.github.bobi.aemgroovyconsoleplugin.services.http
 import com.github.bobi.aemgroovyconsoleplugin.services.PasswordsService
 import com.github.bobi.aemgroovyconsoleplugin.services.http.model.GroovyConsoleOutput
 import com.github.bobi.aemgroovyconsoleplugin.services.http.model.GroovyConsoleTable
+import com.github.bobi.aemgroovyconsoleplugin.services.model.AemDevToken
 import com.github.bobi.aemgroovyconsoleplugin.services.model.AemServerConfig
+import com.github.bobi.aemgroovyconsoleplugin.services.model.AuthType
 import com.google.gson.Gson
 import com.google.gson.GsonBuilder
 import com.intellij.openapi.Disposable
@@ -11,88 +13,111 @@ import com.intellij.openapi.components.Service
 import com.intellij.openapi.components.service
 import com.intellij.openapi.diagnostic.thisLogger
 import com.intellij.openapi.project.Project
-import org.apache.http.HttpException
-import org.apache.http.HttpResponse
-import org.apache.http.HttpStatus
-import org.apache.http.auth.AuthScope
-import org.apache.http.auth.Credentials
-import org.apache.http.auth.UsernamePasswordCredentials
-import org.apache.http.client.config.RequestConfig
-import org.apache.http.client.methods.HttpPost
-import org.apache.http.client.protocol.HttpClientContext
-import org.apache.http.client.utils.URIBuilder
-import org.apache.http.client.utils.URIUtils
-import org.apache.http.entity.mime.MultipartEntityBuilder
-import org.apache.http.impl.auth.BasicScheme
-import org.apache.http.impl.client.BasicAuthCache
-import org.apache.http.impl.client.BasicCredentialsProvider
-import org.apache.http.impl.client.CloseableHttpClient
-import org.apache.http.impl.client.HttpClientBuilder
+import io.ktor.client.*
+import io.ktor.client.call.*
+import io.ktor.client.engine.cio.*
+import io.ktor.client.plugins.auth.*
+import io.ktor.client.plugins.auth.providers.*
+import io.ktor.client.plugins.contentnegotiation.*
+import io.ktor.client.request.*
+import io.ktor.client.request.forms.*
+import io.ktor.client.statement.*
+import io.ktor.http.*
+import io.ktor.serialization.kotlinx.json.*
+import io.ktor.utils.io.core.*
+import org.apache.hc.client5.http.classic.methods.HttpPost
+import org.apache.hc.client5.http.config.ConnectionConfig
+import org.apache.hc.client5.http.config.RequestConfig
+import org.apache.hc.client5.http.entity.mime.MultipartEntityBuilder
+import org.apache.hc.client5.http.impl.classic.CloseableHttpClient
+import org.apache.hc.client5.http.impl.classic.HttpClients
+import org.apache.hc.client5.http.impl.io.PoolingHttpClientConnectionManagerBuilder
+import org.apache.hc.client5.http.utils.Base64
+import org.apache.hc.core5.http.ClassicHttpResponse
+import org.apache.hc.core5.http.Header
+import org.apache.hc.core5.http.HttpException
+import org.apache.hc.core5.http.HttpStatus
+import org.apache.hc.core5.http.message.BasicHeader
+import org.apache.hc.core5.net.URIBuilder
+import org.apache.hc.core5.util.Timeout
+import java.io.IOException
 import java.io.InputStreamReader
 import java.io.StringReader
 import java.net.URI
-import java.util.concurrent.TimeUnit
+import java.time.Duration
+import kotlin.io.use
+import kotlin.text.toByteArray
 
 /**
  * User: Andrey Bardashevsky
  * Date/Time: 02.08.2022 16:06
  */
 @Service(Service.Level.PROJECT)
-class GroovyConsoleHttpService : Disposable {
-    private val httpClient: CloseableHttpClient = createHttpRequest()
+class GroovyConsoleHttpService(private val project: Project) : Disposable {
 
-    private val gson: Gson = GsonBuilder().create()
+    private val imsTokenProvider = AdobeIMSTokenProvider.getInstance(project)
 
-    private fun createHttpRequest(): CloseableHttpClient {
-        val timeout = TimeUnit.MINUTES.toMillis(10).toInt()
+    private val httpClient: CloseableHttpClient by lazy {
+        val timeout = Timeout.of(Duration.ofMinutes(10))
 
-        return HttpClientBuilder.create()
+        return@lazy HttpClients.custom()
+            .setConnectionManager(
+                PoolingHttpClientConnectionManagerBuilder.create()
+                    .setDefaultConnectionConfig(
+                        ConnectionConfig.custom()
+                            .setConnectTimeout(timeout)
+                            .setSocketTimeout(timeout)
+                            .build()
+                    )
+                    .build()
+            )
             .setDefaultRequestConfig(
                 RequestConfig.custom()
-                    .setConnectTimeout(timeout)
                     .setConnectionRequestTimeout(timeout)
-                    .setSocketTimeout(timeout)
                     .build()
             )
             .build()
     }
 
-    fun execute(url: String, user: String, password: String, script: ByteArray, action: Action) =
-        execute(createAemHttpConfig(url, user, password), script, action)
+    private val gson: Gson = GsonBuilder().create()
 
-    fun execute(config: AemServerConfig, script: ByteArray, action: Action) = execute(createAemHttpConfig(config), script, action)
+    fun execute(httpConfig: AemServerHttpConfig, script: ByteArray) =
+        execute(httpConfig, script, Action.EXECUTE)
+
+    fun execute(config: AemServerConfig, script: ByteArray, action: Action) =
+        execute(
+            AemServerHttpConfig(
+                id = config.id,
+                url = config.url,
+                authType = config.authType,
+                credentials = getCredentials(config)
+            ),
+            script,
+            action
+        )
 
     private fun execute(config: AemServerHttpConfig, script: ByteArray, action: Action): GroovyConsoleOutput {
-        val uri = URIBuilder().also {
+        val uri = URIBuilder().apply {
             val configHostUri = URI.create(config.url)
 
-            it.scheme = configHostUri.scheme
-            it.host = configHostUri.host
-            it.port = configHostUri.port
+            scheme = configHostUri.scheme
+            host = configHostUri.host
+            port = configHostUri.port
 
-            it.path = action.path
+            path = action.path
         }.build()
 
-        val httpHost = URIUtils.extractHost(uri)
+        val request = HttpPost(uri).apply {
+            addHeader(getAuthorizationHeader(config))
 
-        val context: HttpClientContext = HttpClientContext.create().also { context ->
-            context.authCache = BasicAuthCache().also { cache ->
-                cache.put(httpHost, BasicScheme())
-            }
-            context.credentialsProvider = BasicCredentialsProvider().also { credentialsProvider ->
-                credentialsProvider.setCredentials(AuthScope(httpHost), config.credentials)
-            }
+            entity = MultipartEntityBuilder.create().addBinaryBody("script", script).build()
         }
 
-        val request = HttpPost(uri).also {
-            it.entity = MultipartEntityBuilder.create().addBinaryBody("script", script).build()
-        }
-
-        return httpClient.execute(httpHost, request, ::handleResponse, context)
+        return httpClient.execute(request, ::handleResponse)
     }
 
-    private fun handleResponse(response: HttpResponse): GroovyConsoleOutput {
-        if (HttpStatus.SC_OK == response.statusLine.statusCode) {
+    private fun handleResponse(response: ClassicHttpResponse): GroovyConsoleOutput {
+        if (HttpStatus.SC_OK == response.code) {
             val output = InputStreamReader(response.entity.content).use {
                 return@use gson.fromJson(it, Output::class.java)
             }
@@ -120,8 +145,17 @@ class GroovyConsoleHttpService : Disposable {
                 table = outputTable?.table
             )
         } else {
-            throw HttpException(response.statusLine.toString())
+            throw HttpException("${response.code} ${response.reasonPhrase}")
         }
+    }
+
+    private fun getAuthorizationHeader(config: AemServerHttpConfig): Header {
+        val value = when (config.authType) {
+            AuthType.BASIC -> "Basic " + String(Base64.encodeBase64("${config.credentials.user}:${config.credentials.password}".toByteArray()))
+            else -> "Bearer " + imsTokenProvider.getAccessToken(config)
+        }
+
+        return BasicHeader("Authorization", value)
     }
 
     override fun dispose() {
@@ -132,7 +166,7 @@ class GroovyConsoleHttpService : Disposable {
 
         fun getInstance(project: Project): GroovyConsoleHttpService = project.service()
 
-        private fun getCredentials(config: AemServerConfig): Credentials {
+        private fun getCredentials(config: AemServerConfig): AemServerCredentials {
             val credentials = PasswordsService.getCredentials(config.id)
 
             val user = credentials?.userName.orEmpty()
@@ -142,26 +176,31 @@ class GroovyConsoleHttpService : Disposable {
                 throw IllegalArgumentException("Credentials is not found for server: ${config.id}")
             }
 
-            return UsernamePasswordCredentials(user, password)
-        }
-
-        private fun createAemHttpConfig(cfg: AemServerConfig): AemServerHttpConfig {
-            return AemServerHttpConfig(url = cfg.url, credentials = getCredentials(cfg))
-        }
-
-        private fun createAemHttpConfig(url: String, user: String, password: String): AemServerHttpConfig {
-            return AemServerHttpConfig(url = url, credentials = UsernamePasswordCredentials(user, password))
+            return AemServerCredentials(user, password)
         }
     }
 
-    private data class Output(val output: String? = null, val runningTime: String? = null, val exceptionStackTrace: String? = null, val result: String? = null)
+    private data class Output(
+        val output: String? = null,
+        val runningTime: String? = null,
+        val exceptionStackTrace: String? = null,
+        val result: String? = null
+    )
 
     private data class OutputTable(val table: GroovyConsoleTable)
-
-    private class AemServerHttpConfig(val url: String, val credentials: Credentials)
 
     enum class Action(internal val path: String) {
         EXECUTE("/bin/groovyconsole/post.json"),
         EXECUTE_DISTRIBUTED("/bin/groovyconsole/replicate"),
     }
 }
+
+class AemServerCredentials(val user: String, val password: String)
+
+class AemServerHttpConfig(
+    val id: Long? = null,
+    val url: String,
+    val authType: AuthType,
+    val credentials: AemServerCredentials
+)
+